adjusted chat controller

+++ backend/src/main/java/com/nicolas/chatapp/dto/request/SingleChatRequestDTO.java
@@ -1,4 +1,4 @@
 package com.nicolas.chatapp.dto.request;
 
-public record SingleChatRequestDTO(Integer userId) {
+public record SingleChatRequestDTO(Long userId) {
 }
